create a utils.ts file to avoid circular reference
typescript can handle the circular ref but it just feels bad to have one. index.ts was getting big anyway
